evm: move more eof files
