+++ packages/evm/src/eof/constants.ts
@@ -1,0 +1,9 @@
+// Constants
+export const FORMAT = 0xef
+
+export const MAGIC = Buffer.from('EF00', 'hex')
+
+export const KIND_TYPE = 1
+export const KIND_CODE = 2
+export const KIND_DATA = 3
+export const TERMINATOR = 0
